chore: renamed subfolder of scalismo-faces ply read/writer files
