+++ src/main/scala/scalismo/io/plyScalismoFaces/PlyMeshWriter.scala
@@ -13,16 +13,16 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package scalismo.io.ply
+package scalismo.io.plyScalismoFaces
 
 import java.io.{File, FileOutputStream, OutputStream, OutputStreamWriter}
 import java.nio.ByteOrder
 
 import scalismo.color.RGBA
-import scalismo.io.ply.PlyHelpers.PlyFormat._
-import scalismo.io.ply.PlyHelpers.PlyHeader._
-import scalismo.io.ply.PlyHelpers._
-import scalismo.io.ply.PlyMeshPropertyWriters._
+import scalismo.io.plyScalismoFaces.PlyHelpers.PlyFormat._
+import scalismo.io.plyScalismoFaces.PlyHelpers.PlyHeader._
+import scalismo.io.plyScalismoFaces.PlyHelpers._
+import scalismo.io.plyScalismoFaces.PlyMeshPropertyWriters._
 import scalismo.geometry._
 import scalismo.mesh._
