+++ src/main/scala/scalismo/io/plyScalismoFaces/Readers.scala
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package scalismo.io.ply
+package scalismo.io.plyScalismoFaces
 
 import java.io.{IOException, InputStream}
 import java.nio._
