+++ src/main/scala/scalismo/io/plyScalismoFaces/Writers.scala
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package scalismo.io.ply
+package scalismo.io.plyScalismoFaces
 
 import java.io.{OutputStream, OutputStreamWriter}
 import java.nio.{ByteBuffer, ByteOrder}
